backend/requirements.txt
# Core Django and Web Framework
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
python-dotenv>=1.0.0

# Database Support
django-redis>=5.3.0     # Redis cache backend

# Authentication and Security
# Token authentication is built into djangorestframework

# FastAPI and async support
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0          # For async HTTP requests
python-multipart>=0.0.6  # For form data and file uploads

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Task Queue (Background Processing)
celery>=5.3.0
redis>=5.0.0

# Production Server Support
gunicorn>=21.0.0       # WSGI server for production
whitenoise>=6.5.0      # Static file serving

# Monitoring and Logging
sentry-sdk[django]>=1.32.0  # Error tracking (optional)

# Development and Testing
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.3.0
factory-boy>=3.3.0
freezegun>=1.2.0

# API Documentation
drf-spectacular>=0.26.0  # OpenAPI 3 schema generation

# File Upload and Processing
Pillow>=10.0.0         # Image processing
openpyxl>=3.1.0        # Excel file support

# Validation and Serialization
marshmallow>=3.20.0
pydantic>=2.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3 